+++ src/main/java/es/in2/issuer/application/workflow/impl/VerifiableCredentialIssuanceWorkflowImpl.java
@@ -1,32 +1,23 @@
 package es.in2.issuer.application.workflow.impl;
 
 import com.nimbusds.jose.JWSObject;
-import com.upokecenter.cbor.CBORObject;
 import es.in2.issuer.application.workflow.VerifiableCredentialIssuanceWorkflow;
-import es.in2.issuer.domain.exception.Base45Exception;
 import es.in2.issuer.domain.exception.InvalidOrMissingProofException;
+import es.in2.issuer.domain.exception.ParseErrorException;
 import es.in2.issuer.domain.model.dto.*;
-import es.in2.issuer.domain.model.enums.SignatureType;
 import es.in2.issuer.domain.service.*;
+import es.in2.issuer.domain.util.factory.LEARCredentialEmployeeFactory;
 import es.in2.issuer.infrastructure.config.AppConfig;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import nl.minvws.encoding.Base45;
-import org.apache.commons.compress.compressors.CompressorOutputStream;
-import org.apache.commons.compress.compressors.CompressorStreamFactory;
 import org.springframework.stereotype.Service;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
-import reactor.core.scheduler.Schedulers;
 
-import java.io.ByteArrayOutputStream;
 import java.text.ParseException;
-import java.util.Base64;
-import java.util.Collections;
 import java.util.UUID;
 
-import static es.in2.issuer.domain.util.Constants.CWT_VC;
-import static es.in2.issuer.domain.util.Constants.JWT_VC;
+import static es.in2.issuer.domain.util.Constants.*;
 
 @Slf4j
 @Service
@@ -41,15 +32,18 @@ public class VerifiableCredentialIssuanceWorkflowImpl implements VerifiableCrede
     private final EmailService emailService;
     private final CredentialProcedureService credentialProcedureService;
     private final DeferredCredentialMetadataService deferredCredentialMetadataService;
+    private final TrustFrameworkService trustFrameworkService;
+    private final LEARCredentialEmployeeFactory credentialEmployeeFactory;
+
 
     @Override
-    public Mono<Void> completeWithdrawLearCredentialProcess(String processId, String type, LEARCredentialRequest learCredentialRequest) {
-        return verifiableCredentialService.generateVc(processId, type, learCredentialRequest)
+    public Mono<Void> completeIssuanceCredentialProcess(String processId, String type, IssuanceRequest issuanceRequest) {
+        return verifiableCredentialService.generateVc(processId, type, issuanceRequest)
                 .flatMap(transactionCode -> {
-                    String email = learCredentialRequest.credential().get("mandatee").get("email").asText();
-                    String firstName =  learCredentialRequest.credential().get("mandatee").get("first_name").asText();
-                    return emailService.sendTransactionCodeForCredentialOffer(email, "Credential Offer", appConfig.getIssuerUiExternalDomain() + "/credential-offer?transaction_code=" + transactionCode, firstName,appConfig.getWalletUrl());                });
-
+                    String email = issuanceRequest.payload().get("mandatee").get("email").asText();
+                    String firstName = issuanceRequest.payload().get("mandatee").get("first_name").asText();
+                    return emailService.sendTransactionCodeForCredentialOffer(email, "Credential Offer", appConfig.getIssuerUiExternalDomain() + "/credential-offer?transaction_code=" + transactionCode, firstName, appConfig.getWalletUrl());
+                });
     }
 
     @Override
@@ -59,25 +53,48 @@ public class VerifiableCredentialIssuanceWorkflowImpl implements VerifiableCrede
             String token
     ) {
         try {
-        JWSObject jwsObject = JWSObject.parse(token);
-        String authServerNonce = jwsObject.getPayload().toJSONObject().get("jti").toString();
-        return proofValidationService.isProofValid(credentialRequest.proof().jwt(), token)
-                .flatMap(isValid -> {
-                    if (Boolean.FALSE.equals(isValid)) {
-                        return Mono.error(new InvalidOrMissingProofException("Invalid proof"));
-                    } else {
-                        return extractDidFromJwtProof(credentialRequest.proof().jwt());
-                    }
-                })
-                .flatMap(subjectDid -> verifiableCredentialService.buildCredentialResponse(processId, subjectDid, authServerNonce, credentialRequest.format())
-                        .flatMap(credentialResponse -> deferredCredentialMetadataService.getProcedureIdByAuthServerNonce(authServerNonce)
-                                .flatMap(credentialProcedureService::getSignerEmailFromDecodedCredentialByProcedureId)
-                                .flatMap(email -> emailService.sendPendingCredentialNotification(email,"Pending Credential")
-                                        .then(Mono.just(credentialResponse)))));
-        }
-        catch (ParseException e){
+            JWSObject jwsObject = JWSObject.parse(token);
+            String authServerNonce = jwsObject.getPayload().toJSONObject().get("jti").toString();
+
+            return proofValidationService.isProofValid(credentialRequest.proof().jwt(), token)
+                    .flatMap(isValid -> {
+                        if (Boolean.FALSE.equals(isValid)) {
+                            return Mono.error(new InvalidOrMissingProofException("Invalid proof"));
+                        } else {
+                            return extractDidFromJwtProof(credentialRequest.proof().jwt());
+                        }
+                    })
+                    .flatMap(subjectDid ->
+                                deferredCredentialMetadataService.getOperationModeByAuthServerNonce(authServerNonce)
+                                        .flatMap(operationMode ->
+                                                verifiableCredentialService.buildCredentialResponse(processId, subjectDid, authServerNonce, credentialRequest.format(), token, operationMode)
+                                                        .flatMap(credentialResponse -> {
+                                                                    if (operationMode.equals(ASYNC)) {
+                                                                        return deferredCredentialMetadataService.getProcedureIdByAuthServerNonce(authServerNonce)
+                                                                                .flatMap(credentialProcedureService::getSignerEmailFromDecodedCredentialByProcedureId)
+                                                                                .flatMap(email ->
+                                                                                        emailService.sendPendingCredentialNotification(email, "Pending Credential")
+                                                                                                .then(Mono.just(credentialResponse))
+                                                                                );
+                                                                    } else if (operationMode.equals(SYNC)) {
+                                                                        return deferredCredentialMetadataService.getProcedureIdByAuthServerNonce(authServerNonce)
+                                                                                .flatMap(id -> credentialProcedureService.updateCredentialProcedureCredentialStatusToValidByProcedureId(id)
+                                                                                        .then(credentialProcedureService.getDecodedCredentialByProcedureId(id)
+                                                                                                .flatMap(decodedCredential -> processDecodedCredential(processId,decodedCredential))
+                                                                                        )
+                                                                                )
+                                                                                .then(deferredCredentialMetadataService.deleteDeferredCredentialMetadataByAuthServerNonce(authServerNonce))
+                                                                                .then(Mono.just(credentialResponse));
+                                                                    } else {
+                                                                        return Mono.error(new IllegalArgumentException("Unknown operation mode: " + operationMode));
+                                                                    }
+                                                                }
+                                                        )
+                                        )
+                    );
+        } catch (ParseException e) {
             log.error("Error parsing the accessToken", e);
-            throw new RuntimeException("Error parsing accessToken", e);
+            throw new ParseErrorException("Error parsing accessToken");
         }
     }
 
@@ -102,7 +119,7 @@ public class VerifiableCredentialIssuanceWorkflowImpl implements VerifiableCrede
 
     @Override
     public Mono<VerifiableCredentialResponse> generateVerifiableCredentialDeferredResponse(String processId, DeferredCredentialRequest deferredCredentialRequest) {
-                return verifiableCredentialService.generateDeferredCredentialResponse(processId,deferredCredentialRequest)
+        return verifiableCredentialService.generateDeferredCredentialResponse(processId, deferredCredentialRequest)
                 .onErrorResume(e -> Mono.error(new RuntimeException("Failed to process the credential for the next processId: " + processId, e)));
     }
 
@@ -123,78 +140,6 @@ public class VerifiableCredentialIssuanceWorkflowImpl implements VerifiableCrede
 //                    .onErrorResume(e -> Mono.error(new RuntimeException("Failed to sign and update the credential.", e)));
     }
 
-    @Override
-    public Mono<String> signCredentialOnRequestedFormat(String unsignedCredential, String format, String userId, UUID credentialId, String token) {
-        return Mono.defer(() -> {
-            if (format.equals(JWT_VC)) {
-                log.info(unsignedCredential);
-                log.info("Signing credential in JADES remotely ...");
-                SignatureRequest signatureRequest = new SignatureRequest(
-                        new SignatureConfiguration(SignatureType.JADES, Collections.emptyMap()),
-                        unsignedCredential
-                );
-                return remoteSignatureService.sign(signatureRequest, token)
-                        .publishOn(Schedulers.boundedElastic())
-                        .map(SignedData::data);
-            } else if (format.equals(CWT_VC)) {
-                log.info(unsignedCredential);
-                return generateCborFromJson(unsignedCredential)
-                        .flatMap(cbor -> generateCOSEBytesFromCBOR(cbor, token))
-                        .flatMap(this::compressAndConvertToBase45FromCOSE);
-            } else {
-                return Mono.error(new IllegalArgumentException("Unsupported credential format: " + format));
-            }
-        });
-    }
-
-    /**
-     * Generate CBOR payload for COSE.
-     *
-     * @param edgcJson EDGC payload as JSON string
-     * @return Mono emitting CBOR bytes
-     */
-    private Mono<byte[]> generateCborFromJson(String edgcJson) {
-        return Mono.fromCallable(() -> CBORObject.FromJSONString(edgcJson).EncodeToBytes());
-    }
-
-    /**
-     * Generate COSE bytes from CBOR bytes.
-     *
-     * @param cbor  CBOR bytes
-     * @param token Authentication token
-     * @return Mono emitting COSE bytes
-     */
-    private Mono<byte[]> generateCOSEBytesFromCBOR(byte[] cbor, String token) {
-        log.info("Signing credential in COSE format remotely ...");
-        String cborBase64 = Base64.getEncoder().encodeToString(cbor);
-        SignatureRequest signatureRequest = new SignatureRequest(
-                new SignatureConfiguration(SignatureType.COSE, Collections.emptyMap()),
-                cborBase64
-        );
-        return remoteSignatureService.sign(signatureRequest, token).map(signedData -> Base64.getDecoder().decode(signedData.data()));
-    }
-
-    /**
-     * Compress COSE bytes and convert it to Base45.
-     *
-     * @param cose COSE Bytes
-     * @return Mono emitting COSE bytes compressed and in Base45
-     */
-    private Mono<String> compressAndConvertToBase45FromCOSE(byte[] cose) {
-        return Mono.fromCallable(() -> {
-            ByteArrayOutputStream stream = new ByteArrayOutputStream();
-            try (CompressorOutputStream deflateOut = new CompressorStreamFactory()
-                    .createCompressorOutputStream(CompressorStreamFactory.DEFLATE, stream)) {
-                deflateOut.write(cose);
-            } // Automatically closed by try-with-resources
-            byte[] zip = stream.toByteArray();
-            return Base45.getEncoder().encodeToString(zip);
-        }).onErrorResume(e -> {
-            log.error("Error compressing and converting to Base45: " + e.getMessage(), e);
-            return Mono.error(new Base45Exception("Error compressing and converting to Base45"));
-        });
-    }
-
     private Mono<String> extractDidFromJwtProof(String jwtProof) {
         return Mono.fromCallable(() -> {
             JWSObject jwsObject = JWSObject.parse(jwtProof);
@@ -204,5 +149,47 @@ public class VerifiableCredentialIssuanceWorkflowImpl implements VerifiableCrede
             return kid.split("#")[0];
         });
     }
+
+
+    private Mono<Void> processDecodedCredential(String processId, String decodedCredential) {
+        log.info("ProcessID: {} Decoded Credential: {}", processId, decodedCredential);
+
+        LEARCredentialEmployeeJwtPayload learCredentialEmployeeJwtPayload = credentialEmployeeFactory.mapStringToLEARCredentialEmployee(decodedCredential);
+
+        String signerOrgIdentifier = learCredentialEmployeeJwtPayload.learCredentialEmployee().credentialSubject().mandate().signer().organizationIdentifier();
+        if (signerOrgIdentifier == null || signerOrgIdentifier.isBlank() ){
+            log.error("ProcessID: {} Signer Organization Identifier connot be null or empty", processId);
+            return Mono.error(new IllegalArgumentException("Organization Identifier not valid"));
+        }
+
+        String mandatorOrgIdentifier = learCredentialEmployeeJwtPayload.learCredentialEmployee().credentialSubject().mandate().mandator().organizationIdentifier();
+        if (mandatorOrgIdentifier == null || mandatorOrgIdentifier.isBlank() ){
+            log.error("ProcessID: {} Mandator Organization Identifier connot be null or empty", processId);
+            return Mono.error(new IllegalArgumentException("Organization Identifier not valid"));
+        }
+
+        return saveToTrustFramework(processId,signerOrgIdentifier, mandatorOrgIdentifier);
+    }
+
+    private Mono<Void> saveToTrustFramework(String processId, String signerOrgIdentifier, String mandatorOrgIdentifier) {
+
+        String signerDid = DID_ELSI + signerOrgIdentifier;
+        String mandatorDid = DID_ELSI + mandatorOrgIdentifier;
+
+        return trustFrameworkService.validateDidFormat(processId, signerDid)
+                .flatMap(isValid -> registerDidIfValid(processId, signerDid, isValid))
+                .then(trustFrameworkService.validateDidFormat(processId, mandatorDid)
+                        .flatMap(isValid -> registerDidIfValid(processId, mandatorDid,isValid)));
+    }
+
+    private Mono<Void> registerDidIfValid(String processId, String did, boolean isValid) {
+        if (isValid) {
+            return trustFrameworkService.registerDid(processId, did);
+        } else {
+            log.error("ProcessID: {} Did not registered because is invalid", processId);
+            return Mono.empty();
+        }
+    }
+
 
 }