+++ src/main/java/es/in2/issuer/domain/service/VerifiableCredentialService.java
@@ -1,15 +1,13 @@
 package es.in2.issuer.domain.service;
 
 import es.in2.issuer.domain.model.dto.DeferredCredentialRequest;
-import es.in2.issuer.domain.model.dto.LEARCredentialRequest;
+import es.in2.issuer.domain.model.dto.IssuanceRequest;
 import es.in2.issuer.domain.model.dto.VerifiableCredentialResponse;
 import reactor.core.publisher.Mono;
 
 public interface VerifiableCredentialService {
-    Mono<String> generateVc(String processId, String vcType, LEARCredentialRequest learCredentialRequest);
-//    Mono<String> generateVcPayLoad(String vcTemplate, String subjectDid, String issuerDid, String userData, Instant expiration);
-    Mono<VerifiableCredentialResponse> buildCredentialResponse(String processId, String subjectDid, String authServerNonce, String format);
-
+    Mono<String> generateVc(String processId, String vcType, IssuanceRequest issuanceRequest);
+    Mono<VerifiableCredentialResponse> buildCredentialResponse(String processId, String subjectDid, String authServerNonce, String format, String token, String operationMode);
     Mono<Void> bindAccessTokenByPreAuthorizedCode(String processId, String accessToken, String preAuthCode);
     Mono<VerifiableCredentialResponse> generateDeferredCredentialResponse(String processId, DeferredCredentialRequest deferredCredentialRequest);
 }