+++ src/main/java/es/in2/issuer/infrastructure/controller/CredentialSignerController.java
@@ -6,9 +6,10 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
 import reactor.core.publisher.Mono;
+
+import static es.in2.issuer.domain.util.Constants.JWT_VC;
 
 @Slf4j
 @RestController
@@ -23,6 +24,6 @@ public class CredentialSignerController {
     public Mono<Void> createVerifiableCredential(
             @RequestHeader(HttpHeaders.AUTHORIZATION) String authorizationHeader,
             @RequestBody ProcedureIdRequest procedureIdRequest) {
-        return credentialSignerWorkflow.signCredential(authorizationHeader, procedureIdRequest.procedureId());
+        return credentialSignerWorkflow.signAndUpdateCredentialByProcedureId(authorizationHeader, procedureIdRequest.procedureId(), JWT_VC).then();
     }
 }