+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v1.1.0
+### Added
+- LEARCredentialEmployee issuance in a synchronous way.
+- DOME Trust Framework integration to register issuers and participants.
+### Changed
+- Issuances API to support various issuance types.
+
 ## [Unreleased]: v0.7.0
 - LEARCredential compliance.
 
@@ -23,14 +30,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Batch credential support (extra)
 
 ## [Unreleased]: v0.3.0
-- Support for credential in JWT and CWT.
+- Support for credentials in JWT and CWT.
 - Remove of external libraries for CV generation
 - Native credential payload generation.
 - Local emission.
 
 ## [Unreleased]: v0.2.0
 - Adapter for Abstract Configuration loading.
-- Support for Configurations from Yaml file.
+- Support for Configurations from YAML file.
 - Support for Configurations from Azure App Configuration.
 
 ## [Unreleased]: v0.1.0